feat: fp32 master-weight guard, EMA regression tests, dtype-safe checkpoint restore
Adds the safety net for the mixed-precision effort: a UserWarning guard in
AbstractPipeline.__init__ against non-fp32 master weights (the bf16-EMA bug
this whole plan exists to fix), EMA regression tests documenting why bf16
master weights break optax.ema/AdamW moment accumulation, and dtype-safe
orbax checkpoint restore (casts restored leaves to the target model's
dtypes, mirroring load_safetensors' existing astype loop).

Also fixes get_default_flux1joint_params, which still hard-coded
param_dtype=jnp.bfloat16 (a leftover pre-mixed-precision default missed by
the earlier per-model migration) -- surfaced immediately by the new guard
firing on the default-constructed Flux1Joint pipelines.

+++ src/gensbi/recipes/flux1joint.py
@@ -79,7 +79,7 @@ def get_default_flux1joint_params(dim_joint: int, in_channels: int = 1):
         dim_joint=dim_joint,
         id_merge_mode="concat",
         guidance_embed=False,
-        param_dtype=jnp.bfloat16,
+        param_dtype=jnp.float32,
     )
